fix(frontend): add loader component and add loading for topic & legislator page
